backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
snowflake-connector-python==3.7.0
snowflake-snowpark-python==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
